+++ docs/versionhistory.rst
@@ -8,6 +8,11 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added support for the ``from_uri()``, ``full_match()``, ``parser`` methods/properties
   in ``anyio.Path``, newly added in Python 3.13
   (`#737 <https://github.com/agronholm/anyio/issues/737>`_)
+- Added support for more keyword arguments for ``run_process()`` and ``open_process()``:
+  ``preexec_fn``, ``startupinfo``, ``creationflags``, ``user``, ``group``,
+  ``extra_groups`` and ``umask``
+- Improved the type annotations in ``run_process()`` and ``open_process()`` to allow for
+  path-like arguments, just like ``subprocess.Popen``
 - Changed the ``ResourceWarning`` from an unclosed memory object stream to include its
   address for easier identification
 - Fixed ``to_process.run_sync()`` failing to initialize if ``__main__.__file__`` pointed